Small file rename client side
