+++ client-vue/src/backend/settings.service.ts
@@ -1,7 +1,7 @@
 import { ServerApi } from "@/backend/server.api";
 import { BaseService } from "@/backend/base.service";
 import { PrinterFileCleanSubSetting, ServerSettings } from "@/models/server-settings/server-settings.model";
-import { PrinterFileCleanSettings } from "@/models/server-settings/file-handling-settings.model";
+import { PrinterFileCleanSettings } from "@/models/server-settings/printer-file-clean-settings.model";
 import { ClientSettings } from "@/models/client-settings/client-settings.model";
 
 export class SettingsService extends BaseService {
